:sparkles: Add hal::runtime and hal::unsafe
These initializer tags help guide the writer and reader to the kinds
operations they are asking their system to perform. It makes clear
that an operation will have to perform runtime checks and if an
operation is inherently unsafe without being very certain that the
input parameters are correct.

Document initializers.