:ambulance: Add on_bus_off API to hal::buffered_can
This API should have been in the original `hal::can` interface. I had
forgotten that bus-off also means that reception is disabled meaning,
using only the `send()` API to communicate that the system is "bus-off"
means requiring code to use that API to figure that out, which is bad
form to busy the bus with noop messages just to determine bus off state.
And rather than making developers poll this information, we can use an
interrupt.