:sparkles: Add [strong/weak/optional]_ptr
These smart pointers will be the bedrock of libhal libraries, ensuring
that the lifetimes of objects used by drivers stays for as long as those
drivers are active. We are migrating to allocator aware dynamic memory
model for libhal. We plan to provide safe and deterministic allocators
such as arenas (for drivers) and stack-like allocators (for coroutines)
backed by static memory which makes it possible to account for in the
memory budget.

- hal::strong_ptr<T> is our light weight version of C++'s `std::shared_ptr`,
  with an atomic reference count. It is never non-null or invalid.
  Even a moved from strong_ptr will be valid to access until it is no
  longer accessible by code. If you have a strong_ptr its valid memory
- hal::optional_ptr<T> is a nullable version of strong_ptr<T>. If you access
  it and it is null (or invalid), it will throw a
  `std::bad_optional_access();`
- hal::weak_ptr<T> is near identical to std::weak_ptr but for our smart
  pointers. When you `lock()` a weak_ptr() it returns an optional_ptr.