:truck: Rename interrupt classes (#144)
- rename timer to v5::timed_interrupt
- rename interrupt_pin to v5::edge_triggered_interrupt
- rename can_interrupt to v5::can_message_interrupt

The purpose of this is to make the names of the interrupt classes
consistent and explanatory. Now when spoken developers can speak of
these things as interrupt types. We may say, "Do we need a timed
interrupt or an can_message interrupt. Can we rely on the can messages
to ensure we wake up often enough? Or do we need something there to
periodically wake up the system to check for work to be done."

`hal::adc_conversion_complete_interrupt` may exist but most likely we'd
make an `adc_interrupt` interface with the necessary number of callback
installing functions.