Add `sync_wait` to async_context
I realized this morning that if we wanted to distribute work in an
io_uring type fashion, we'd need a consistent type that can be used
to make forward progress on the coroutine stack. async_context is such
an object. With the `sync_wait`, a async_context can be passed to
an API and its result can be waited for until its complete. Another
addition should be a `resume()` function.

The sync_wait function also reduces the number of comparisons a
coroutine operation takes. Now, we use whether or not the handle within
the context becomes noop.