:zap: Make async<T> -> future<T>
Future types are types that can by executed to get an answer, just
like with std::future. But instead we are using coroutines where you
can resume steps or sync_wait the whole operation. When the coroutine
has completed, then the future's value is changed from the previous
coroutine into the result of the coroutine. The coroutine that generated
the value destroy's itself after setting the value.

The allows for promise types to be smaller since we do not have to
store the whole type T within the promise. This also allows functions
that can complete synchronously to return back a value directly rather
than perform the coroutine initialization if its not necessary.

I'm considering extending the promise type to support co_yield which
will set the value of the future, but will not destroy itself. This
would allow our promise type to support being a generator allowing
for pre-allocated coroutines.