:tada: Fully working coroutines
What we have currently should work for the needs of the libhal. When
we feel comfortable enough to migrate the v5 APIs over to coroutines,
then we shall move the coroutine.hpp file out of experimental.

I've dropped the concept of a generator for now as I think its too
much of an early optimization. We should work on engineering it when
we have a direct and needed use case.