:sparkles: Add `hal::v5::scatter_span`
Scatter span will be the defacto-standard for all APIs that deal
with arrays of data in v5. At its core it is a span-of-spans with type
`std::span<std::span<T> const>`. This allows code to order disjoint
non-contiguous regions of memory and pass all of them to an API to be
transmitted or filled.

**Motivation:**

Previously this required calling APIs for each disjoint memory block OR
allocating a region of memory to buffer the disjoint memory regions into
a single contiguous block of memory. The first option means that the
code will have to perform multiple calls to an API which can result
time gaps between each segment of memory (transmitted or received).
The second allocates additional memory the size of all blocks of memory
(or in chunk) in order to perform this operation.

Rather than copy data or perform multiple calls, we can allocate an
array of spans (in general should be smaller), preferably on the stack,
and send that to API.