✨ (minor) Add usb_interface to abstract USB device implementations (#163)
* :sparkles: Implemented a high-level USB interface class made to represent USB features outside of base configuration.
For example, an interface could be a speaker, keyboard, mouse, or serial device. This is intended to be used at the level below the configuration descriptor. Which can either be an interface descriptor or an interface assoication descriptor. This interface is expected to manage its own endpoints and interact with the control endpoint (EP0) via callbacks.
usb.hpp which previously had the endpoint declarations now is just a shortcut for including interface.hpp and endpoint.hpp. The endpoints are now in endpoint.hpp

* :white_check_mark: test(usb_interface): Wrote test for single iface descriptor iface

* fix(usb_interface): Set interface takes a u8 not u16

* :clown_face: test(usb_interface): Mocked an IAD
based interface for testing

* Slimming down the interfaces and updating names

* Remove usb_interfaces.test.cpp

* Rename setup -> setup_packet

---------

Co-authored-by: Khalil Estell <kammce0@gmail.com>