➕ Add strong_ptr & async_context to v5 (#180)
This change removes `hal::callback` as it will be replaced with an `mem::strong_ptr<CallbackInterface>` or `mem::optional_ptr<CallbackInterface>`. In all cases where we'd want a callback function, we will provide an interface that contains an API that can be invoked as the callback. This way, we we have:

1. Control over the allocation of callbacks via PMR, unlike with `std::function`.
2. Can have callbacks that are not constrained by an arbitrary size, like `stdex::inplace_function`.
3. The lifetime of the object is maintained and managed